final version

+++ lib/model/timesheet_model.dart
@@ -3,9 +3,12 @@
 
 import 'dart:convert';
 
-List<TimesheetModel> timesheetModelFromJson(String str) => List<TimesheetModel>.from(json.decode(str).map((x) => TimesheetModel.fromJson(x)));
+List<TimesheetModel> timesheetModelFromJson(String str) =>
+    List<TimesheetModel>.from(
+        json.decode(str).map((x) => TimesheetModel.fromJson(x)));
 
-String timesheetModelToJson(List<TimesheetModel> data) => json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
+String timesheetModelToJson(List<TimesheetModel> data) =>
+    json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
 
 class TimesheetModel {
   TimesheetModel({
@@ -22,29 +25,29 @@ class TimesheetModel {
   String weekday;
   String starttime;
   String endtime;
-  int workinghours;
+  double workinghours;
   Project project;
   User user;
 
   factory TimesheetModel.fromJson(Map<String, dynamic> json) => TimesheetModel(
-    tid: json["tid"],
-    weekday: json["weekday"],
-    starttime: json["starttime"],
-    endtime: json["endtime"],
-    workinghours: json["workinghours"],
-    project: Project.fromJson(json["project"]),
-    user: User.fromJson(json["user"]),
-  );
+        tid: json["tid"],
+        weekday: json["weekday"],
+        starttime: json["starttime"],
+        endtime: json["endtime"],
+        workinghours: json["workinghours"],
+        project: Project.fromJson(json["project"]),
+        user: User.fromJson(json["user"]),
+      );
 
   Map<String, dynamic> toJson() => {
-    "tid": tid,
-    "weekday": weekday,
-    "starttime": starttime,
-    "endtime": endtime,
-    "workinghours": workinghours,
-    "project": project.toJson(),
-    "user": user.toJson(),
-  };
+        "tid": tid,
+        "weekday": weekday,
+        "starttime": starttime,
+        "endtime": endtime,
+        "workinghours": workinghours,
+        "project": project.toJson(),
+        "user": user.toJson(),
+      };
 }
 
 class Project {
@@ -57,14 +60,14 @@ class Project {
   String name;
 
   factory Project.fromJson(Map<String, dynamic> json) => Project(
-    pid: json["pid"],
-    name: json["name"],
-  );
+        pid: json["pid"],
+        name: json["name"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "pid": pid,
-    "name": name,
-  };
+        "pid": pid,
+        "name": name,
+      };
 }
 
 class User {
@@ -83,18 +86,18 @@ class User {
   String email;
 
   factory User.fromJson(Map<String, dynamic> json) => User(
-    uid: json["uid"],
-    username: json["username"],
-    firstname: json["firstname"],
-    lastname: json["lastname"],
-    email: json["email"],
-  );
+        uid: json["uid"],
+        username: json["username"],
+        firstname: json["firstname"],
+        lastname: json["lastname"],
+        email: json["email"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "uid": uid,
-    "username": username,
-    "firstname": firstname,
-    "lastname": lastname,
-    "email": email,
-  };
+        "uid": uid,
+        "username": username,
+        "firstname": firstname,
+        "lastname": lastname,
+        "email": email,
+      };
 }
